integrate neuralforecast
