+++ requirements.txt
@@ -1,13 +1,13 @@
 rpy2
 sktime==0.16.1
 gluonts==0.12.3
-pytorch-lightning==1.9.4
 mxnet==1.9.1
 prophet
 kaleido
 Jinja2
 omegaconf
 statsforecast==1.4.0
+neuralforecast
 missingno
 lightgbm
 tbats==1.1.2
